backend/database/auth.py
from passlib.context import CryptContext
from jose import jwt
from datetime import datetime, timedelta
from backend.config import SUPER_SECRET_KEY

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
SECRET_KEY = SUPER_SECRET_KEY
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30


def hash_password(password: str):
    return pwd_context.hash(password)


def verify_password(plain_password, hashed_password):
    return pwd_context.verify(plain_password, hashed_password)


def create_access_token(data: dict):
    to_encode = data.copy()
    expire = datetime.now(datetime.timezone.utc) + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
