requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.23.1
python-dotenv==1.0.0
google-genai
black==25.1.0
flake8==7.2.0
starlette==0.27.0
pydantic-ai==0.2.12
python-jose==3.5.0
passlib==1.7.4
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
pydantic[email]
passlib[argon2]